requirements.txt
numpy==1.23.3
pillow==9.2.0